refactor: organize hooks and features
