+++ src/hooks/useCustomers.js
@@ -1,0 +1,141 @@
+import { PHASES, RECIPES } from '../constants';
+
+const useCustomers = (gameState, setGameState, addEvent, addNotification, setSelectedCustomer) => {
+  const generateCustomers = () => {
+    const customerCount = Math.floor(Math.random() * 4) + 3;
+    const customers = [];
+
+    const getRarityWeights = (day) => {
+      if (day <= 3) {
+        return { common: 70, uncommon: 30, rare: 0 };
+      } else if (day <= 6) {
+        return { common: 50, uncommon: 40, rare: 10 };
+      } else {
+        return { common: 30, uncommon: 45, rare: 25 };
+      }
+    };
+
+    const getRandomRarity = (weights) => {
+      const rand = Math.random() * 100;
+      let threshold = 0;
+      for (const [rarity, weight] of Object.entries(weights)) {
+        threshold += weight;
+        if (rand <= threshold) return rarity;
+      }
+      return 'common';
+    };
+
+    for (let i = 0; i < customerCount; i++) {
+      const requests = ['weapon', 'armor', 'trinket'];
+      const rarityWeights = getRarityWeights(gameState.day);
+
+      const requestType = requests[Math.floor(Math.random() * requests.length)];
+      const requestRarity = getRandomRarity(rarityWeights);
+      const basePrice = requestRarity === 'common' ? 15 : requestRarity === 'uncommon' ? 25 : 50;
+      const offerPrice = Math.floor(basePrice * (0.9 + Math.random() * 0.3));
+
+      const flexibility = Math.random();
+      const isFlexible = flexibility > 0.6;
+
+      customers.push({
+        id: i,
+        name: `Customer ${i + 1}`,
+        requestType,
+        requestRarity,
+        offerPrice,
+        satisfied: false,
+        isFlexible,
+        patience: Math.floor(Math.random() * 3) + 2
+      });
+    }
+
+    return customers;
+  };
+
+  const openShop = () => {
+    const customers = generateCustomers();
+    setGameState(prev => ({
+      ...prev,
+      phase: PHASES.SHOPPING,
+      customers
+    }));
+    addEvent(`Shop opened with ${customers.length} customers waiting`, 'info');
+  };
+
+  const serveCustomer = (customerId, itemId) => {
+    const customer = gameState.customers.find(c => c.id === customerId);
+    const recipe = RECIPES.find(r => r.id === itemId);
+
+    if (!customer || !recipe || (gameState.inventory[itemId] || 0) < 1) return;
+
+    let payment = customer.offerPrice;
+    let satisfaction = 'perfect';
+
+    const exactMatch = recipe.type === customer.requestType && recipe.rarity === customer.requestRarity;
+
+    if (!exactMatch) {
+      let penalty = 0.4;
+
+      if (customer.isFlexible) {
+        penalty = 0.2;
+        satisfaction = 'good substitute';
+      } else {
+        satisfaction = 'reluctant';
+      }
+
+      const rarityOrder = { common: 1, uncommon: 2, rare: 3 };
+      if (rarityOrder[recipe.rarity] > rarityOrder[customer.requestRarity]) {
+        penalty -= 0.1;
+        satisfaction = customer.isFlexible ? 'delighted upgrade' : 'acceptable upgrade';
+      }
+
+      if (recipe.type === customer.requestType) {
+        penalty -= 0.1;
+      }
+
+      payment = Math.floor(payment * (1 - penalty));
+    }
+
+    const newInventory = { ...gameState.inventory };
+    newInventory[itemId] -= 1;
+
+    const newCustomers = gameState.customers.map(c =>
+      c.id === customerId ? { ...c, satisfied: true, payment, satisfaction } : c
+    );
+
+    setGameState(prev => ({
+      ...prev,
+      inventory: newInventory,
+      customers: newCustomers,
+      gold: prev.gold + payment,
+      totalEarnings: prev.totalEarnings + payment
+    }));
+
+    const matchText = exactMatch ? '(Perfect match!)' : `(${satisfaction})`;
+    addEvent(`Sold ${recipe.name} to ${customer.name} for ${payment} gold ${matchText}`, 'success');
+    addNotification(`💰 Sold ${recipe.name} for ${payment} gold!`, 'success');
+    setSelectedCustomer(null);
+  };
+
+  const endDay = () => {
+    setGameState(prev => ({
+      ...prev,
+      phase: PHASES.END_DAY
+    }));
+  };
+
+  const startNewDay = () => {
+    setGameState(prev => ({
+      ...prev,
+      phase: PHASES.MORNING,
+      day: prev.day + 1,
+      customers: []
+    }));
+    addEvent(`Started Day ${gameState.day + 1}`, 'info');
+    setSelectedCustomer(null);
+  };
+
+  return { openShop, serveCustomer, endDay, startNewDay };
+};
+
+export default useCustomers;
